// src/websites/komorii/src/types/api.ts
export interface ManifestFile {
    hash: string;
    filename: string;
    size: number;
}

export interface FullChapterManifest {
    version: string;
    seriesId: string;
    chapterId: string;
    chapterNumber: number; // or ensure type match
    files: ManifestFile[];
    nodes?: string[];
}